dom/base/test/test_window_enumeration.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=807222
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 807222</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=807222">Mozilla Bug 807222</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 807222 **/
var expectedProps = [ "Image", "Audio", "Option",
                      "PerformanceTiming", "CSS2Properties", "SVGElement" ];
if (navigator.mozTelephony) {
  expectedProps.push("USSDReceivedEvent");
}
var actualProps = Object.getOwnPropertyNames(window);

for (var i = 0; i < expectedProps.length; ++i) {
  isnot(actualProps.indexOf(expectedProps[i]), -1,
        "getOwnPropertyNames should include " + expectedProps[i]);
}
</script>
</pre>
</body>
</html>
